Add component directory and restructure files
